Adding user logged verification step also inside the backend

+++ src/main/kotlin/Routing.kt
@@ -21,6 +21,9 @@ fun Route.projectRoutes(repos: Repositories) {
         }
 
         get("/{id}") {
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
             val projectID = call.parameters["id"]?.toIntOrNull()
             if (projectID == null) {
                 call.respond(HttpStatusCode.BadRequest, "ID must be an integer")
@@ -30,6 +33,13 @@ fun Route.projectRoutes(repos: Repositories) {
             val project = repos.projectRepository.getProjectByID(projectID)
             if (project == null) {
                 call.respond(HttpStatusCode.NotFound, "Project not found")
+                return@get
+            }
+
+            if (project.owner.id != userId &&
+                project.collaborators.find { collaborator -> collaborator.id == userId } == null) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this project")
                 return@get
             }
 
@@ -48,7 +58,21 @@ fun Route.projectRoutes(repos: Repositories) {
         }
 
         put {
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
             val entry = call.receive<ProjectEntryRequest>()
+
+            val project = repos.projectRepository.getProjectByID(entry.id ?: 0)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Project not found")
+                return@put
+            }
+
+            if (project.owner.id != userId) {
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this project")
+                return@put
+            }
 
             val edited = repos.projectRepository.editProject(entry)
             if (edited == null) {
@@ -66,6 +90,20 @@ fun Route.projectRoutes(repos: Repositories) {
                 return@delete
             }
 
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val project = repos.projectRepository.getProjectByID(id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Project not found")
+                return@delete
+            }
+
+            if (project.owner.id != userId) {
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this project")
+                return@delete
+            }
+
             if (!repos.projectRepository.deleteProject(id)) {
                 call.respond(HttpStatusCode.NotFound, "Unable to remove project")
                 return@delete
@@ -80,6 +118,21 @@ fun Route.collaboratorRoutes(repos: Repositories) {
     route("/collaborators") {
         post {
             val collaborator = call.receive<CollaboratorEntryRequest>()
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val project = repos.projectRepository.getProjectByID(collaborator.projectID)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Project not found")
+                return@post
+            }
+
+            if (project.owner.id != userId) {
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this project")
+                return@post
+            }
+
             val created = repos.projectRepository.addCollaboratorToProject(collaborator)
             if (created == null) {
                 call.respond(HttpStatusCode.InternalServerError, "Failed to add collaborator to project")
@@ -91,6 +144,21 @@ fun Route.collaboratorRoutes(repos: Repositories) {
 
         delete {
             val collaborator = call.receive<CollaboratorEntryRequest>()
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val project = repos.projectRepository.getProjectByID(collaborator.projectID)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Project not found")
+                return@delete
+            }
+
+            if (project.owner.id != userId) {
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this project")
+                return@delete
+            }
+
             if (!repos.projectRepository.removeCollaboratorToProject(collaborator)) {
                 call.respond(HttpStatusCode.NotFound, "Project not found")
                 return@delete
@@ -103,12 +171,29 @@ fun Route.collaboratorRoutes(repos: Repositories) {
 
 fun Route.issueRoutes(repos: Repositories) {
     route("/issues") {
-        get {
-            call.respond(repos.issueRepository.getAllIssues())
-        }
+        // get {
+        //     call.respond(repos.issueRepository.getAllIssues())
+        // }
 
         post {
             val issue = call.receive<IssueEntryRequest>()
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val project = repos.projectRepository.getProjectByID(issue.projectID ?: 0)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Project not found")
+                return@post
+            }
+
+            if (project.owner.id != userId &&
+                project.collaborators.find { user -> user.id == userId } == null) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this project")
+                return@post
+            }
+
             val created = repos.issueRepository.createIssue(issue)
             if (created == null) {
                 call.respond(HttpStatusCode.InternalServerError, "Failed to create issue")
@@ -131,11 +216,49 @@ fun Route.issueRoutes(repos: Repositories) {
                 return@get
             }
 
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val project = repos.projectRepository.getProjectByID(issue.project.id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue's project not found")
+                return@get
+            }
+
+            if (project.owner.id != userId &&
+                project.collaborators.find { user -> user.id == userId } == null) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this issue")
+                return@get
+            }
+
             call.respond(issue)
         }
 
         put {
             val entry = call.receive<IssueEntryRequest>()
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val issue = repos.issueRepository.getIssueByID(entry.id ?: 0)
+            if (issue == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue not found")
+                return@put
+            }
+
+            val project = repos.projectRepository.getProjectByID(issue.project.id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue's project not found")
+                return@put
+            }
+
+            if (project.owner.id != userId &&
+                issue.createdBy.id != userId) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this issue")
+                return@put
+            }
 
             val edited = repos.issueRepository.editIssue(entry)
             if (edited == null) {
@@ -150,6 +273,28 @@ fun Route.issueRoutes(repos: Repositories) {
             val id = call.parameters["id"]?.toIntOrNull()
             if (id == null) {
                 call.respond(HttpStatusCode.BadRequest, "ID must be an integer")
+                return@delete
+            }
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val issue = repos.issueRepository.getIssueByID(id)
+            if (issue == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue not found")
+                return@delete
+            }
+
+            val project = repos.projectRepository.getProjectByID(issue.project.id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue's project not found")
+                return@delete
+            }
+
+            if (project.owner.id != userId &&
+                issue.createdBy.id != userId) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this issue")
                 return@delete
             }
 
@@ -168,6 +313,39 @@ fun Route.linkRoutes(repos: Repositories) {
         post {
             val link = call.receive<IssueLinkEntryRequest>()
 
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val linker_issue = repos.issueRepository.getIssueByID(link.linker)
+            if (linker_issue == null) {
+                call.respond(HttpStatusCode.NotFound, "Linker issue not found")
+                return@post
+            }
+
+            val linked_issue = repos.issueRepository.getIssueByID(link.linked)
+            if (linked_issue == null) {
+                call.respond(HttpStatusCode.NotFound, "Linked issue not found")
+                return@post
+            }
+
+            val project = repos.projectRepository.getProjectByID(linker_issue.project.id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Linker issue's project not found")
+                return@post
+            }
+
+            if (project.owner.id != userId &&
+                project.collaborators.find { user -> user.id == userId } == null) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this issue")
+                return@post
+            }
+
+            if (linker_issue.project.id != linked_issue.project.id) {
+                call.respond(HttpStatusCode.BadRequest, "Only issues under the same project can be linked")
+                return@post
+            }
+
             val created = repos.issueRepository.linkIssues(link)
             if (created == null) {
                 call.respond(HttpStatusCode.InternalServerError, "Failed to create link")
@@ -179,6 +357,29 @@ fun Route.linkRoutes(repos: Repositories) {
 
         delete {
             val link = call.receive<IssueLinkEntryRequest>()
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val issue = repos.issueRepository.getIssueByID(link.linker)
+            if (issue == null) {
+                call.respond(HttpStatusCode.NotFound, "Linker issue not found")
+                return@delete
+            }
+
+            val project = repos.projectRepository.getProjectByID(issue.project.id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Linker issue's project not found")
+                return@delete
+            }
+
+            if (project.owner.id != userId &&
+                issue.createdBy.id != userId) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this issue")
+                return@delete
+            }
+
             if (!repos.issueRepository.deleteLink(link)) {
                 call.respond(HttpStatusCode.NotFound, "Unable to remove link")
                 return@delete
@@ -194,6 +395,28 @@ fun Route.commentRoutes(repos: Repositories) {
         post {
             val comment = call.receive<CommentEntryRequest>()
 
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val issue = repos.issueRepository.getIssueByID(comment.issueID ?: 0)
+            if (issue == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue not found")
+                return@post
+            }
+
+            val project = repos.projectRepository.getProjectByID(issue.project.id)
+            if (project == null) {
+                call.respond(HttpStatusCode.NotFound, "Issue's project not found")
+                return@post
+            }
+
+            if (project.owner.id != userId &&
+                project.collaborators.find { user -> user.id == userId } == null) {
+
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this issue")
+                return@post
+            }
+
             val created = repos.commentRepository.createComment(comment)
             if (created == null) {
                 call.respond(HttpStatusCode.InternalServerError, "Failed to create comment")
@@ -204,9 +427,23 @@ fun Route.commentRoutes(repos: Repositories) {
         }
 
         put {
-            val comment = call.receive<CommentEntryRequest>()
+            val entry = call.receive<CommentEntryRequest>()
 
-            val edited = repos.commentRepository.editComment(comment)
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val comment = repos.commentRepository.getCommentByID(entry.id ?: 0)
+            if (comment == null) {
+                call.respond(HttpStatusCode.NotFound, "Comment not found")
+                return@put
+            }
+
+            if (comment.author.id != userId) {
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this comment")
+                return@put
+            }
+
+            val edited = repos.commentRepository.editComment(entry)
             if (edited == null) {
                 call.respond(HttpStatusCode.NotFound, "Comment requested not found")
                 return@put
@@ -219,6 +456,21 @@ fun Route.commentRoutes(repos: Repositories) {
             val commentID = call.parameters["id"]?.toIntOrNull()
             if (commentID == null) {
                 call.respond(HttpStatusCode.BadRequest, "ID must be an integer")
+                return@delete
+            }
+
+            val principal = call.principal<JWTPrincipal>()
+            val userId = principal!!.payload.getClaim("userId").asInt()
+
+            val comment = repos.commentRepository.getCommentByID(commentID)
+            if (comment == null) {
+                call.respond(HttpStatusCode.NotFound, "Comment not found")
+                return@delete
+            }
+
+            // TODO: technically also the project owner and issue owner will be able to delete the comment
+            if (comment.author.id != userId) {
+                call.respond(HttpStatusCode.Unauthorized, "Not authorized to access this comment")
                 return@delete
             }
 
@@ -238,9 +490,8 @@ fun Route.userRoutes(repos: Repositories) {
         //     call.respond(repos.userRepository.getAllUsers())
         // }
 
-        post {
-
-        }
+        // post {
+        // }
 
         // get("/{id}") {
         //     val uid = call.parameters["id"]?.toIntOrNull()
